requirements.txt
pygame==2.5.2
numpy==1.26.4
ollama==0.1.7
requests==2.31.0
chromadb==0.4.22
openai==1.12.0
anthropic==0.7.8
python-dotenv==1.0.0